refactor: mv src/gui/src/HotkeyDialog => src/gui/src/dialogs/HotkeyDialog
