+++ src/gui/src/dialogs/HotkeyDialog.cpp
