+++ src/gui/src/dialogs/HotkeyDialog.h
