+++ src/gui/src/dialogs/HotkeyDialog.ui
